refactor(bench): share fresh-module probe runner (#8972)
Extract the isolate_roots sweep into a declarative reference/candidate harness. Rotate pair order, alternate orientation, preserve signed candidate-minus-reference deltas, enforce per-module axiom policies, and reject transitive imports between measured modules. Keep the real-roots suite as a thin adapter and run the shared regression tests in the existing CI job.\n\nSession handoff: progress/20260728T002605Z_fresh-module-harness.md
